+++ nTA/Source/net_SendBuffer.cpp
@@ -131,7 +131,7 @@ void net_SendBuffer::SendTo( net_Client& target )
 void net_SendBuffer::CompletePacket()
 {
 	if( (m_WriteCursor - m_WorkingPacket)>sizeof(net_PacketHeader) )
-		((net_PacketHeader*)m_WorkingPacket)->Size = m_WriteCursor - m_WorkingPacket,
+		((net_PacketHeader*)m_WorkingPacket)->Size = (UINT32)(m_WriteCursor - m_WorkingPacket),
 		theApp.Console.Comment( CT_DEBUG, "net_SendBuffer::CompletePacket(): %d byte packet to %d completed", ((net_PacketHeader*)m_WorkingPacket)->Size, ((net_PacketHeader*)m_WorkingPacket)->Marker ),
 		m_WorkingPacket = m_WriteCursor,
 		((net_PacketHeader*)m_WorkingPacket)->Marker = ~1;
@@ -189,7 +189,7 @@ bool net_SendBuffer::NextPacket( net_Client& target )
 //
 void net_SendBuffer::NewPacket( UINT32 uiTarget, UINT32 uiSize )
 {
-	((net_PacketHeader*)m_WorkingPacket)->Size = m_WriteCursor - m_WorkingPacket;
+	((net_PacketHeader*)m_WorkingPacket)->Size = (UINT32)(m_WriteCursor - m_WorkingPacket);
 	theApp.Console.Comment( CT_DEBUG, "net_SendBuffer::NewPacket( %d, %u ); Old(%d,%u)", uiTarget, uiSize, ((net_PacketHeader*)m_WorkingPacket)->Marker, ((net_PacketHeader*)m_WorkingPacket)->Size ),
 	m_WorkingPacket = m_WriteCursor;
